requirements.txt
aiogram>=3.27,<4
SQLAlchemy
psycopg2-binary
emoji
python-dotenv
alembic==1.17.2
pyrogram
tgcrypto
tqdm
APScheduler>=3.10
tzdata
aiohttp>=3.9
openai>=1.40
numpy>=1.24
scikit-learn>=1.4
pgvector>=0.3